+++ docs/docs/CHANGELOG.md
@@ -46,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Significantly improve the performances of the PoW validation
 - Add trimpath option to artifact builds
 - Add COOKIE_HTTP_ONLY option to set the HttpOnly flag on Anubis cookies
+- Only compute the JA4H fingerprint when a policy references the `X-Http-Fingerprint-JA4H` header, taking it off the hot path for configurations that don't use it ([#834](https://github.com/TecharoHQ/anubis/pull/834)).
 
 ## v1.25.0: Necron